requirements.txt
pandas==2.2.3
openpyxl==3.1.5

python-docx==1.1.2

python-barcode==0.15.1
Pillow==10.4.0

pywin32==306

pyinstaller==6.5.0